Adding `empty?` Predicate to HL7 API
There are often times when consumers need to know whether the path
they are selecting from the HL7 contains data. This concept of the
data returned being "empty" is currently undefinded in the HL7
library and is left up to interpretation of the users.

This provides an implementaiton for determining if a given path
is empty of data or not.